requirements.txt
pandas
requests
yfinance
python-telegram-bot==20.6
